refactor: Add cn utility function for class name merging
The code changes in this commit add a new utility function called `cn` to the `utils.ts` file. This function is responsible for merging multiple class names using the `clsx` and `twMerge` libraries. It takes in an array of class values and returns the merged class name. This utility function will be used for handling class names in the project.

Note: The commit message follows the established conventions of the repository.